Expose WithTooltip / cleanup styles / add documenation for tooltipable components (#424)
