+++ CHANGELOG.md
@@ -56,6 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added ZoomIn and ZoomOut icons
 - Added style variants to Caution icon
 - Added WithTooltip() HOC and updated RosterHeader, RosterCell, PopOverMenu, PopOver, NavbarItem, ControlbarItem, and ChatBubbleConatiner to support tooltips
+- Added documentation for components that support tooltips, exposed WithTooltip component and related interfaces/types
 
 ### Changed
