+++ src/components/ui/Chat/ChatBubble/ChatBubble.mdx
@@ -11,6 +11,8 @@ import { GlobalStyles } from '../../../../theme/GlobalStyles';
 The ChatBubble component displays incoming and outgoing chat messages and data.
 It also determines the style of the bubble including color, chat bubble tail visibility, and name visibility.
 
+ChatBubbles support tooltips on any buttons by adding a `data-tooltip` property. Adding a `data-tooltip-position` property controls the tooltip position
+
 ## Importing
 
 ```javascript
