+++ src/components/ui/ControlBar/ControlBar.mdx
@@ -15,6 +15,8 @@ You can place ControlBar into the parent element according to the `layout` prop,
 
 The ControlBar component contains one or more `ControlBarItem` components, which can show a label and icon. You can add a dropdown list to each ControlBarItem by setting the [popOver props](/?path=/docs/ui-components-popover--basic-pop-over-menu).
 
+ControlBars support tooltips on any buttons by adding a `data-tooltip` property. Adding a `data-tooltip-position` property controls the tooltip position
+
 ## Importing
 
 ```javascript
