+++ src/components/ui/Navbar/Navbar.mdx
@@ -14,6 +14,8 @@ import { GlobalStyles } from '../../../theme/GlobalStyles';
 
 The Navbar component displays a vertical menu with navigation icon buttons and labels. Navbar is constructed using the NavbarHeader and NavbarItem components.
 
+Navbars support tooltips on any buttons by adding a `data-tooltip` property. Adding a `data-tooltip-position` property controls the tooltip position
+
 ## Importing
 
 ```javascript
