+++ src/components/ui/Roster/Roster.mdx
@@ -14,6 +14,8 @@ import { GlobalStyles } from '../../../theme/GlobalStyles';
 The Roster component displays all the attendees in a meeting including their local statuses for audio, video, content share and if running late.
 It is made up of one RosterHeader component and multiple RosterGroup components.
 
+RosterHeaders and RosterCells support tooltips on any buttons by adding a `data-tooltip` property. Adding a `data-tooltip-position` property controls the tooltip position
+
 ## Importing
 
 ```javascript
